requirements.txt
streamlit
requests
plotly
python-dotenv
python-dotenv
